Resolve unnecessary_cast clippy lint
    error: casting to the same type is unnecessary (`i32` -> `i32`)
      --> src/log.rs:51:5
       |
    51 |     ((e * c - s) >> SHIFT_AMOUNT) as i32
       |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `((e * c - s) >> SHIFT_AMOUNT)`
       |
       = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
       = note: `-D clippy::unnecessary-cast` implied by `-D clippy::all`

    error: casting to the same type is unnecessary (`u32` -> `u32`)
       --> src/to_chars.rs:118:20
        |
    118 |     let left_end = (((1 << (MAX_PRECISION + ADDITIONAL_PRECISION)) + DIVISOR - 1) / DIVISOR) as u32;
        |                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `(((1 << (MAX_PRECISION + ADDITIONAL_PRECISION)) + DIVISOR - 1) / DIVISOR)`
        |
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast

    error: casting to the same type is unnecessary (`u32` -> `u32`)
       --> src/to_chars.rs:119:21
        |
    119 |     let right_end = (((1 << ADDITIONAL_PRECISION) * ((1 << MAX_PRECISION) + 1)) / DIVISOR) as u32;
        |                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `(((1 << ADDITIONAL_PRECISION) * ((1 << MAX_PRECISION) + 1)) / DIVISOR)`
        |
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast

+++ src/log.rs
@@ -48,7 +48,7 @@ const fn compute<
     //debug_assert!(e <= MAX_EXPONENT && e >= -MAX_EXPONENT);
     let c = floor_shift(C_INTEGER_PART, C_FRACTIONAL_DIGITS, SHIFT_AMOUNT);
     let s = floor_shift(S_INTEGER_PART, S_FRACTIONAL_DIGITS, SHIFT_AMOUNT);
-    ((e * c - s) >> SHIFT_AMOUNT) as i32
+    (e * c - s) >> SHIFT_AMOUNT
 }
 
 const LOG10_2_FRACTIONAL_DIGITS: u64 = 0x4d10_4d42_7de7_fbcc;
